Reinstantiate windows rather than mutating `StateObject`s; use immutability
Some code (which is deleted prior to this commit) utilised methods which made `StateObject`s interiorly mutable by objects which did not own them, resulting in their unnecessary reinstantiation. Part of the solution was used in the last commit: to reinstantiate the windows as required, rather than the objects, to adhere to the mutability enforced by structs and for all observed objects to remain immutable but interioly mutable by the view which owns them.